- Added DebugMacros.cpp and DebugMacros.h files - Renamed variables for better understandability - Moved some code from Core.h to DebugMacros files - Updated CMakeLists.txt to include new files - Made minor formatting and syntax corrections across multiple files
More detailed explanation:
This commit adds Debug Macros for console trace, log, and error messages, assertions and profiling to new DebugMacros files. Renamed some variables for better understanding, and ensured correct use of `const` qualifiers. Also included new files in CMakeLists for project building and made minor syntax and formatting changes across multiple files. This helps in the better organization of the code and improves overall code readability.

This commit message was generated by AI.
