+++ src/Ayla/aypch.h
@@ -26,6 +26,7 @@
 
 // Other Stuff
 #include "Ayla/Core/Core.h"
+#include "Ayla/Core/DebugMacros.h"
 
 
 #ifdef AYLA_PLATFORM_WINDOWS
